+++ src/components/styled-atlaskit/DynamicTable/DynamicTable.stories.tsx
@@ -45,4 +45,6 @@ Example.args = {
       ],
     },
   ],
+  children: <p>Some text below the table</p>,
+  customContent: <span>Some optional custom content above the table</span>
 };